+++ plex/myplex_test.go
@@ -1,33 +1,10 @@
 package plex
 
 import (
-	"fmt"
 	"log"
 	"os"
 	"testing"
 )
-
-type fakeCredPrompter struct {
-	username string
-	password string
-}
-
-type fakeTokenRequester struct {
-}
-
-func (fcp fakeCredPrompter) promptCreds() credentials {
-	return credentials{username: fcp.username, password: fcp.password}
-}
-
-func (ftr fakeTokenRequester) tokenRequest(cred credentials) (string, error) {
-	if cred.username == "ValidUser" {
-		return "ValidToken", nil
-	} else if cred.username == "BadUser" {
-		return "BadToken", fmt.Errorf("badtoken")
-	} else {
-		return "", fmt.Errorf("unknown")
-	}
-}
 
 // Test the failure to create a token file.
 func TestInvalidTokenFile(t *testing.T) {
@@ -36,17 +13,8 @@ func TestInvalidTokenFile(t *testing.T) {
 	tokenFile = "zzz:/invalidPath/tokenfile"
 	defer func() { tokenFile = oldtokenfile }()
 
-	// Fake the credentials returned.
-	fc := fakeCredPrompter{
-		username: "TestUser",
-		password: "TestPass",
-	}
-
-	// Fake the token request.
-	ft := fakeTokenRequester{}
-
 	// Check if the token function returns the value from the test token file.
-	_, err := Token(fc, ft)
+	err := cacheToken("test")
 	if err == nil {
 		t.Error("Was able to create invalid token file")
 	}
@@ -66,17 +34,8 @@ func TestTokenFileRead(t *testing.T) {
 	}
 	f.WriteString("ValidToken")
 
-	// Fake the credentials returned.
-	fc := fakeCredPrompter{
-		username: "TestUser",
-		password: "TestPass",
-	}
-
-	// Fake the token request.
-	ft := fakeTokenRequester{}
-
 	// Check if the token function returns the value from the test token file.
-	token, _ := Token(fc, ft)
+	token := Token()
 	if token != "ValidToken" {
 		t.Error("Tokenfile does not contain 'ValidToken'")
 	}
@@ -93,17 +52,13 @@ func TestTokenGeneration(t *testing.T) {
 	tokenFile = "testTokenFile"
 	defer func() { tokenFile = oldtokenfile }()
 
-	// Fake the credentials returned.
-	fc := fakeCredPrompter{
-		username: "ValidUser",
-		password: "ValidPass",
+	err := cacheToken("ValidToken")
+	if err != nil {
+		t.Error("Unable to generate tokenfile")
 	}
 
-	// Fake the token request.
-	ft := fakeTokenRequester{}
-
 	// Check if the token function returns the value from the test token file.
-	token, _ := Token(fc, ft)
+	token := Token()
 	if token != "ValidToken" {
 		t.Error("Generated token does not contain 'ValidToken'")
 	}
@@ -118,31 +73,3 @@ func TestTokenGeneration(t *testing.T) {
 		log.Printf("Error removing file: %s", err)
 	}
 }
-
-// Test for invalid username/password.
-func TestInvalidCredentials(t *testing.T) {
-	// Replace the tokenfile path for the duration of this test.
-	oldtokenfile := tokenFile
-	tokenFile = "testTokenFile"
-	defer func() { tokenFile = oldtokenfile }()
-
-	// Fake the credentials returned.
-	fc := fakeCredPrompter{
-		username: "Invalid",
-		password: "Invalid",
-	}
-
-	tr := TokenRequester{}
-
-	// Check if the token function returns the value from the test token file.
-	_, err := Token(fc, tr)
-	if err == nil {
-		t.Error("Invalid credentials did not cause error")
-	}
-}
-
-// TODO: Add test for blank token from token file.
-
-// TODO: Add test for no token returned.
-
-// TODO: Add test for no username/password supplied.