+++ CHANGELOG.md
@@ -5,9 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+
+## [v0.2.0]
 ### Added
 - Ability to remove a cached token from the web interface.
 - Logging to file.
+- Logging to web interface.
 - Web interface for management of the application.
 
 ### Changed
@@ -28,5 +32,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - App would attempt to sync even if a token was not obtained.
 - Spaces in TV Show names cause errors.
 
-[unreleased]: https://github.com/danstis/Plex-Sync/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/danstis/Plex-Sync/compare/v0.2.0...HEAD
+[v0.2.0]: https://github.com/danstis/Plex-Sync/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/danstis/Plex-Sync/compare/v0.0.1...v0.1.0